TCA구조 적용, 피드백 적용
