+++ FlowFolio/FlowFolio/Occupation/SelectJobCellData.swift
@@ -1,0 +1,7 @@
+
+// 직군선택 카드뷰에 들어갈 데이터 모델
+struct SelectJobCellData {
+    let title: String
+    let subtitle: String
+    let imageName: String
+}
